Added SOM and On-Page Clustering Output
Added SOM (used by default in mainScript), however this does not seem to
be returning the optimal clusters - need to play around with the
parameters.

Also added clustering of results on the Google page.

I have disabled (by commenting) the code which does this automatically.
To view in action uncomment lines 140-142 in mainScript.js.